+++ full-stack-contactMgmtSys/src/main/java/com/contactmgmtsystem/full_stack_contactMgmtSys/controllers/ProfileController.java
@@ -63,30 +63,6 @@ public class ProfileController {
     }
 
 
-    @RequestMapping(value="/do-register", method = RequestMethod.POST)
-    public String processRegister(@ModelAttribute UserForm userForm){
-        System.out.println("procseeing registraiton");
-/*  Steps to follow for the registration:
- 1). Fetch the data or Input from the form
- 2). Validate the input data
- 3). Save into the database
- 4). Message if needed
- 5). Redirect to login page
-*/
-    User user = User.builder()
-            .name(userForm.getName())
-            .email(userForm.getEmail())
-            .password(userForm.getPassword())
-            .about(userForm.getAbout())
-            .phoneNumber(userForm.getPhoneNumber())
-            .profilePic("https://t4.ftcdn.net/jpg/07/08/47/75/360_F_708477508_DNkzRIsNFgibgCJ6KoTgJjjRZNJD4mb4.jpg")
-            .build();
-
-
-        User savedUser = userServices.saveUser(user);
-        System.out.println("user saved");
-        return "redirect:/register";
-    }
 }
